+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added documentation for the new `terra-table` component.
 
+* Updated
+  * Updated focusable cell test for `terra-data-grid` to account for focusable elements in a hidden container.
+
 ## 1.40.0 - (October 11, 2023)
 
 * Added